requirements.txt
flask==2.0.1
werkzeug==2.0.1
google-generativeai>=0.7.0
python-dotenv==0.19.0
gunicorn==20.1.0
flask-cors==3.0.10
pymongo>=4.0.0
pandas>=1.3.0
